quantum-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum-core"
version = "0.1.0"
description = "Modular quantum computation engine with gyroscopic stabilization"
authors = [
    {name = "Quantum Core Team", email = "quantum@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=1.21.0",
    "torch>=1.12.0",
    "pydantic>=1.9.0",
    "PyYAML>=6.0",
    "scipy>=1.8.0",
    "networkx>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
gpu = [
    "torch[cu118]>=1.12.0",
]
viz = [
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]
data = [
    "pandas>=1.4.0",
    "pyarrow>=8.0.0",
]
all = [
    "quantum-core[dev,gpu,viz,data]"
]

[project.scripts]
quantum-engine = "quantum_core.apps.engine.main:main"

[project.urls]
Homepage = "https://github.com/quantum-core/quantum-core"
Repository = "https://github.com/quantum-core/quantum-core"
Documentation = "https://quantum-core.readthedocs.io"
Issues = "https://github.com/quantum-core/quantum-core/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["quantum_core*"]

[tool.setuptools.package-dir]
quantum_core = "quantum-core"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=quantum_core",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU (deselect with '-m \"not gpu\"')",
    "integration: marks tests as integration tests",
]